+++ docs/changelog.rst
@@ -4,7 +4,9 @@ Changelog
 v0.9.6
 ^^^^^^
 
-* Added Render demo
+* Added a data export button
+* Added an account deletion button
+* Added a Render demo deployment
 
 v0.9.5
 ^^^^^^